refactor(qconnection): hide many fields of Path and Connection
little improvment
